requirements.txt
################################################
# core
################################################
Django==1.11.15


################################################
# general tools
################################################
Markdown==2.6.6
Pillow==5.3.0
analytics==0.6.5
astral==1.2
beautifulsoup4==4.4.1
celery[redis]==4.2.1
cssselect==0.9.1
cssutils==1.0.1
html5lib==0.9999999
jsonfield==1.0.3
#kombu==3.0.35
nydus==0.11.0
oauthlib==1.1.2
openpyxl==2.3.5
premailer==3.0.0
pygeocoder==1.2.5
pyparsing==2.1.5
python-dateutil==2.7.3 # TODO: check compatibility
python-magic==0.4.11
python-mimeparse==1.5.2
pytz==2016.4
requests-oauthlib==0.6.1
requests==2.21.0
simplejson==3.8.2
six==1.11.0
sqlparse==0.1.19
tolerance==0.1.2
bleach==3.0.2
opengraph==0.5
python-decouple==3.1


################################################
# server
################################################
#gunicorn==19.6.0
uwsgi==2.0.17.1
#MySQL-python==1.2.5
psycopg2-binary
redis==2.10.5
dj-database-url==0.5.0
channels==2.1.6
channels-redis==2.3.2

django_redis==4.10.0


################################################
# django modules
################################################
django-absolute==0.3
django-admin-shortcuts==1.2.7
django-appconf==1.0.2
django-authtools==1.5.0
django-click==2.1.0
#django-celery==3.1.17
django-celery-beat==1.4.0
django-classy-tags==0.7.2
django-compressor==2.0
django-countries==3.4.1
django-debug-toolbar==1.4
#django-email-as-username==1.6.7
django-emailit==0.2.2
#django-extensions==2.1.4
django-filer==1.3.2
django-guardian==1.4.4

django-mobile==0.7.0
django-mptt==0.8.7
django-pagedown==1.0.6
django-polymorphic==1.3.1
django-registration==2.2
django-reversion==2.0.2
django-sekizai==0.9.0
django-sendfile==0.3.10
django-settings-export==1.1.0
django-simple-captcha==0.5.9
django-social-auth==0.7.28
django-split-settings==0.3.0 # TODO: remove
django-tastypie==0.13.3
django-xmlrpc==0.1.5
django_braces==1.9.0

# reworking auth / accounts
django-password-reset==2.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0


##################################################################
# RESTful API
##################################################################
djangorestframework==3.9.0


################################################
# django-cms
################################################
django-cms==3.4.6
cmsplugin-filer==1.0.1
cmsplugin-twitter==1.1.0
djangocms-file==1.0
djangocms-gmaps==0.1.4.0
djangocms-link==1.7.2
djangocms-picture==1.0.0
djangocms-snippet==1.8.1
djangocms-teaser==0.2.0
djangocms-text-ckeditor==3.4.0
djangocms-video==1.0.0


################################################
# admin addons
################################################



################################################
# media & audio related
################################################



################################################
# custom repos
################################################
#-e git+https://github.com/hzlf/graypy.git@master#egg=graypy
-e git+https://github.com/palmbeach-interactive/django-slick-admin@master#egg=django-slick-admin

################################################
# services
################################################



################################################
# metrics & insights
################################################
raven==5.0.0






















